Produce: Use libavfilter for conversion, hardware encoding
Use full hardware encoding

Don't download the frame from memory but encode it in-place mapping it with drm-prime and encode with the gpu using vaapi.

When h264_vaapi is asked for but the gpu doesn't actually support encoding h264, smoothly fallback to software encoding.
When a software based encoder is asked for, like libx264 or vp8, completely use the software path as well.