Use the new encoder classes for encoding
Most of the encoder specific code is now in the encoder classes, which
encapsulate encoder specific things. Threading is reworked to have
PipeWireProduce as a pure worker object that we move into a QThread and
then two simpler threads, one to pull from the FFMpeg filtering queue
and pass it onto the encoding queue, the other to pull from the encoding
queue and pass it to PipeWireProduce for processing. Both software and
hardware encoding now rely on FFMpeg to handle queuing.