encodedstream: Add API to set the encoding quality
This allows whoever creates the encodedstream to set the quality level.
It uses a scale of 0%-100% with 0% being worst quality and 100% being
highest quality. This doesn't correspond with values in FFMpeg but those
are rather awkward to handle and even vary per encoder, so instead we
define a standard scale that we convert to an encoder-specific value.