Encoding: Add an environment variable to force encoder type
This adds an environment variable, `KPIPEWIRE_FORCE_ENCODER`, that, when
set to the right value, will force using a certain encoder. Most
importantly, this allows forcing the use of `libx264` software encoding
for H264 if hardware encoding fails for some reason. It also allows
forcing hardware encoding if for some reason aborting when it fails is
preferred.

This is mostly intended as a debugging help.