Remove KPW_WITH_SUGGESTED
There's no situation where KPW_WITH_SUGGESTED is not defined. No code in
this library is actually affected by this definition, so there's no
reason to check if it's defined for users of this library. I can't
imagine a situation where you wouldn't need to know what encoders are
actually available before trying to use one, so there's no reason not to
compile the code that gives a list of suggested encoders.