Add support for the libvpx-vp9 encoder
When it comes to capturing windows being moved around, the smoothness of
libvpx-vp9 is roughly the same as libvpx (VP8) and x264. Moving windows
around looks choppy, but recording a video that is playing on the screen
is perfectly fine. The image is sharp and the compression seems to be
noticeably better than the libvpx encoder. My tests weren't rigorous,
but a libvpx-vp9 video may take up half as much space as a libvpx video
of the same content using the same settings (except for VP9-only
settings). The actual difference depends on the content of the video.
The one thing that's noticeably worse than libvpx is the RAM
consumption. libvpx might make Spectacle consume another ~150MB of RAM,
but libvpx-vp9 will make Spectacle consume ~350MB more than the normal
amount. That could be a problem for low end devices. libvpx-vp9 still
takes up less RAM than x264 with which Spectacle uses ~500MB RAM more
than normal.