Grow PipeWireFrameData to replace QImage to store local frames
While it's handy because QImage can be rendered into, it's useful to us
to be able to pass around the frame data without needing QImage to
support it.
This is necessary for KPipeWire to support processing (both rendering
and recording) frames that aren't RGB textures and the likes, which is
quite common in cameras.
Instead of automatically copying the buffer, hand over to QImage the
function to free the buffer so it can give it away when it's done with.