Replace VaapiUtils::supportsHardwareModifiers with a check on format/modifier
Rather than a binary true/false for all modifiers, use a function that
can check if the current VAAPI can support a certain format and modifier
combination. Right now the implementation of this will only support
linear modifiers since we have no way of actually querying VAAPI but
once VAAPI adds API for this we can replace the implementation here.