Produce: Discard frames if they are arriving faster than max framerate
If the stream fails to reconfigure (for whatever reason) when we specify
a maximum frame rate, we will be receiving frames at a rate higher than
the framerate we want. When an application is using this to throttle how
many frames it can handle, we end up flooding it with too many frames
that the application can't discard since they have already been encoded.
So instead, check the duration between the last and current frame's PTS
and discard the frame if not enough time has elapsed.