Simpler yet more useful handling of KPIPEWIRE_FORCE_ENCODER
Previously, it always overrode the encoder type and profile.

Now just force a specific encoder by inlining the encoder selection
in the switch cases.

This means it's no longer possible to force a different encoder type
than the application requested, but it's arguably not that useful to
e.g. force VP9 if the application expects H.264 packets.