h264vaapi: Use the proper getter for getting the hardware context
The hw_frames_context was moved out of AVFilterLink for FFmpeg 7.1.
Apparently there is actually a proper getter for getting the
hw_frames_ctx of a filter context, so use that instead so things keep
compiling with 7.1.

While at it, add a comment explaining why we need this in the first
place.