encoder: Make it possble to override the filter graph used for software encode
We may want to adjust the filter graph in an encoder to insert
additional filters. Rather than overriding the full method for creating
the filter graph, which would lead to a lot of duplication, this simply
exposes the string that is passed to FFmpeg for parsing.


(cherry picked from commit fa99ca45a5783fb066fc76078bc434551c1d5494)

Co-authored-by: Arjen Hiemstra <ahiemstra@heimr.nl>