Flush encoder queue when streaming doesn't provide an update
Kwin only sends a new frame when there's damage on screen.

The encoder does not flush all frames whilst a stream is active
it will keep one frame in the queue waiting for more input until the
stream is closed.

If there's no update this timer bumps the last frame through the stack
again to flush the last frame.

Testing steps:
- run konsole without any apps blinking cursors, or system monitors
that regularly update.
 - confirm the last pressed key is visible on screen in a timely manner.