Change Encoder::applyEncodingPreference() to buildEncodingOptions()
As it turns out, FFmpeg does something to the pointers here that causes
the dictionary to never be properly used. This means we were never
applying the encoding options, as avcodec_open() was passed a nullptr
for options.

Fix this by changing the function to return an AVDictionary* instead of
trying to modify a passed-in pointer. This results in a proper dict
being returned, that can then be passed to avcodec_open().

The main result of this is that we now properly apply the encoding
options for VP9 encoding, which means we can now encode VP9 at realtime
speeds instead of it massively lagging behind.

BUG: 488896


(cherry picked from commit 52911b7060fc3341c4ab4b308be1e47f985e8576)

Co-authored-by: Arjen Hiemstra <ahiemstra@heimr.nl>