sourcestream: add DRM_FORMAT_MOD_INVALID whether usableModifiers is empty or not.
This commit partially reverts #148. The logic isn't the same as kwin's one.
If a format's modifiers is empty, kwin will not add this format to
result [1].

But in kpipewire part, the format will be added to result whether its
modifiers is empty or not. and empty modifier will cause dmabuf not to
be used [2], this isn't what we expected.

So, to support modifier-less driver (like virtio-gpu) to use
dmabuf instead of memfd, we should add DRM_FORMAT_MOD_INVALID unconditionally.

[1] https://invent.kde.org/plasma/kwin/-/blob/a81e7b749f1a84bb87d53249d5b6a2b53797f578/src/opengl/egldisplay.cpp#L312
[2] https://invent.kde.org/plasma/kpipewire/-/blob/e4b8554cded7b9dd40714ad8d8af3d80bea3fb6a/src/pipewiresourcestream.cpp#L512