Add support for encoding in full color range
Without extra metadata, decoders don't know what color range the frames
of a video use. Most video players will assume correctly that they are
using limited range, except for FFmpeg. This presents a problem for KRdp
as all clients that we support use FFmpeg indirectly for decoding, which
leads to incorrect colors on decode.

To avoid that, add an option to set what color range we should use when
encoding, and use that for the h264 encoders. This allows KRdp to
request full range encoding. The default is still limited range, as this
is also what the upstream default is.

CCBUG: 507015